improve flakiness and potential ids drift
